Refactoring: Make Day -> MarketDay pimp a trait
Done to make the pimping more explicit
