+++ reports/tests/evcel/report/PivotReportTests.scala
@@ -1,29 +1,24 @@
 package evcel.report
 
-import org.scalatest.FunSpec
-import org.scalatest.Matchers
+import org.scalatest.{FunSpec, Matchers}
 import evcel.daterange.DateRangeSugar._
 import evcel.quantity.Qty._
 import evcel.quantity.UOM._
 import evcel.instrument.trade.Trade
 import evcel.report.PivotTrade._
-import evcel.instrument.Future
+import evcel.instrument.{Future, Instrument}
 import scala.language.reflectiveCalls
 import evcel.curve.marketdata.MarketDataTest
-import evcel.curve.environment.MarketDay
-import evcel.curve.environment.MarketDay._
+import evcel.curve.environment.{MarketDay, MarketDayPimps}
 import evcel.curve.UnitTestingEnvironment
-import evcel.valuation.Valuer
-import evcel.valuation.DefaultValuer
-import evcel.instrument.Instrument
-import evcel.pivot.PivotTable
-import evcel.pivot.PivotField
+import evcel.valuation.{Valuer, DefaultValuer}
+import evcel.pivot.{PivotTable, PivotField}
 import evcel.referencedata.market.TestMarkets._
 import evcel.referencedata.market.FuturesMarket
 import evcel.daterange.Month
 import evcel.report.PivotValuer._
 
-class PivotReportTests extends FunSpec with Matchers with MarketDataTest{
+class PivotReportTests extends FunSpec with Matchers with MarketDataTest with MarketDayPimps {
   import PivotReportTests._
   val mkt = "Nymex WTI"
   val marketDay = (1 / Feb / 2014).endOfDay
